Fixed the issue when logout URL has a parameter
When the logout URL has a parameter, the URL was constructed incorrectly, first parameter would be added with a "?" instead of a "&", which resulted in an URL with two "?" which is invalid.

Corrected by checking if the URL already contains a "?", and if it does, add the first parameter with a "&" instead of a "?".

Fixes #134