Use preventClearHashAfterLogin consistently
Firefox will show annoying behavior for setting location.hash to an
empty string if it already is empty: it refreshes the page or scrolls
to the top.

The preventClearHashAfterLogin option is passed down to functions to
prevent this behavior when it's not needed (e.g. during the results
of a silent refresh). However, that option wasn't checked for in a
consistent manner. This commit fixes that.

Fixes #493