fix customHashFragment usage in tryLoginCodeFlow
In tryLoginCodeFlow options.customHashFragment can be passed to override window.location.search in case of a loginFlow running in popup. a querySource const is setup that is assing to that option if set and default to window.location.search but that const was not used to get parts (window.location.search was used instead). This request fix this behavior and correctly use the querySource const.